# academic-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Academic management service for school structure, assignments, and submissions.",
        "title": "Academic Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "paths": {
        "/assignments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List assignments with optional filters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assignments"
                ],
                "summary": "List assignments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class ID",
                        "name": "class_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Subject ID",
                        "name": "subject_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Teacher User ID",
                        "name": "teacher_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Assignment"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Teacher or super admin creates assignment and material.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assignments"
                ],
                "summary": "Create assignment",
                "parameters": [
                    {
                        "description": "Assignment payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Assignment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/classes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List classes with sections and subjects for the authenticated school.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Get classes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ClassWithChildren"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a class for the authenticated school.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Create class",
                "parameters": [
                    {
                        "description": "Class payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateClassRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Class"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sections": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a section under a class.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Create section",
                "parameters": [
                    {
                        "description": "Section payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateSectionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Section"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subjects": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a subject for a class and optional section.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Create subject",
                "parameters": [
                    {
                        "description": "Subject payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateSubjectRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Subject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/submissions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit student work for an assignment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assignments"
                ],
                "summary": "Create submission",
                "parameters": [
                    {
                        "description": "Submission payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateSubmissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Submission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/teacher-assignments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List teacher assignments with optional filters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TeacherAssignments"
                ],
                "summary": "List teacher assignments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Teacher User ID",
                        "name": "teacher_user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Class ID",
                        "name": "class_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Subject ID",
                        "name": "subject_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TeacherAssignment"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assign a teacher user to class and subject.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TeacherAssignments"
                ],
                "summary": "Assign teacher",
                "parameters": [
                    {
                        "description": "Assignment payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateTeacherAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.TeacherAssignment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Assignment": {
            "type": "object",
            "properties": {
                "class_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "material_url": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "teacher_user_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Class": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.ClassWithChildren": {
            "type": "object",
            "properties": {
                "class": {
                    "$ref": "#/definitions/model.Class"
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Section"
                    }
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Subject"
                    }
                }
            }
        },
        "model.CreateAssignmentRequest": {
            "type": "object",
            "required": [
                "class_id",
                "subject_id",
                "title"
            ],
            "properties": {
                "class_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "material_url": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.CreateClassRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.CreateSectionRequest": {
            "type": "object",
            "required": [
                "class_id",
                "name"
            ],
            "properties": {
                "class_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.CreateSubjectRequest": {
            "type": "object",
            "required": [
                "class_id",
                "name"
            ],
            "properties": {
                "class_id": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "section_id": {
                    "type": "string"
                }
            }
        },
        "model.CreateSubmissionRequest": {
            "type": "object",
            "required": [
                "assignment_id",
                "student_id"
            ],
            "properties": {
                "assignment_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "material_url": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                }
            }
        },
        "model.CreateTeacherAssignmentRequest": {
            "type": "object",
            "required": [
                "class_id",
                "subject_id",
                "teacher_user_id"
            ],
            "properties": {
                "class_id": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "teacher_user_id": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
        "model.Section": {
            "type": "object",
            "properties": {
                "class_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Subject": {
            "type": "object",
            "properties": {
                "class_id": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "section_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Submission": {
            "type": "object",
            "properties": {
                "assignment_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "material_url": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "submitted_by": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.TeacherAssignment": {
            "type": "object",
            "properties": {
                "class_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "teacher_user_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}